requirements.txt
boto3
requests
sqlalchemy
snowflake-sqlalchemy
python-dotenv
cryptography
